+++ src/components/PageLayout.js
@@ -4,33 +4,22 @@ import { Fragment } from "react";
 import PropTypes from "prop-types";
 
 import SEO from "./SEO";
-import Header from "./Header";
+import NewHeader from "./NewHeader";
 import Footer from "./Footer";
-import CommonCTA from "./CommonCTA";
-import DividerLine from "./DividerLine";
-import KeepInTheLoop from "./KeepInTheLoop";
 
-const PageLayout = ({
-  title,
-  keywords,
-  children,
-  hideCommonCTA,
-  hideKeepInTheLoop,
-  swoopColor,
-  home
-}) => {
+const PageLayout = ({ title, keywords, children, home }) => {
   return (
     <Fragment>
       <SEO title={title} keywords={keywords} />
-      <Header home={home} />
+      <NewHeader home={home} />
       {children}
-      {!hideCommonCTA && (
+      {/* {!hideCommonCTA && (
         <Fragment>
           <CommonCTA />
           <DividerLine hexColor={swoopColor} />
         </Fragment>
       )}
-      {!hideKeepInTheLoop && <KeepInTheLoop />}
+      {!hideKeepInTheLoop && <KeepInTheLoop />} */}
       <Footer />
     </Fragment>
   );
@@ -43,9 +32,6 @@ PageLayout.propTypes = {
     PropTypes.arrayOf(PropTypes.node),
     PropTypes.node
   ]),
-  hideCommonCTA: PropTypes.bool,
-  hideKeepInTheLoop: PropTypes.bool,
-  swoopColor: PropTypes.string,
   home: PropTypes.bool
 };
